server/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile
#
aniso8601==9.0.1
    # via flask-restful
click==8.0.3
    # via
    #   flask
    #   pip-tools
colorama==0.4.4
    # via click
flask==2.0.2
    # via
    #   -r requirements.in
    #   flask-restful
flask-restful==0.3.9
    # via -r requirements.in
importlib-metadata==4.8.1
    # via
    #   click
    #   pep517
itsdangerous==2.0.1
    # via flask
jinja2==3.0.2
    # via flask
markupsafe==2.0.1
    # via jinja2
pep517==0.12.0
    # via pip-tools
pip-tools==6.4.0
    # via -r requirements.in
pyjwt==1.4.2
    # via -r requirements.in
pytz==2021.3
    # via flask-restful
six==1.16.0
    # via flask-restful
tomli==1.2.1
    # via pep517
typing-extensions==3.10.0.2
    # via importlib-metadata
werkzeug==2.0.2
    # via flask
wheel==0.37.0
    # via pip-tools
zipp==3.6.0
    # via
    #   importlib-metadata
    #   pep517

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
